api/requirements.txt
# pip install -r requirements.txt

fastapi==0.103.1
uvicorn[standard]==0.22.0
pydantic==1.10.9
python-jose
requests
aiofiles

# required for auth
pyjwt[crypto] 
httpx 
python-dotenv